atualizações da pagePost
