+++ lab-eddit/src/CustomHooks/useRequestDataGet.js
@@ -27,46 +27,3 @@ export const useRequestDataGet = (url, initialState) =>{
     return data
 };
 
-export const useRequestDataGetDetail = (url, initialState) =>{
-    const [data, setData] = useState(initialState);
-    const history = useHistory();
-
-    const token = localStorage.getItem('token');
-    if (token === null) {
-        history.push("/");
-    }
-    axios.get(url, {
-        headers:{
-            Authorization: token
-        }
-    }).then((response)=>{
-        console.log(response.data.post)
-        setData(response.data.post)
-    }).catch((error)=>{
-        console.error(error)
-    })
-
-    return data
-}
-export const useDataComments = (url, initialState) =>{
-    const [data, setData] = useState(initialState);
-    const history = useHistory();
-
-    useEffect(()=>{
-        const token = localStorage.getItem('token');
-        if (token === null) {
-            history.push("/");
-        }
-        axios.get(url, {
-            headers:{
-                Authorization: token
-            }
-        }).then((response)=>{
-            console.log(response.data.post.comments)
-            setData(response.data.post.comments)
-        }).catch((error)=>{
-            console.error(error)
-        })
-    },[url, history]);
-    return data
-}
