+++ lab-eddit/src/pages/PagePost/CardComentario.js
@@ -6,13 +6,18 @@ import axios from 'axios';
 import { useParams } from 'react-router-dom';
 
 const CardComentario = (props) => {
-    const [count, setCount] = useState(props.userVoteDirection);
+    let curtida = props.userVoteDirection
     const pathParams = useParams();
     const token = localStorage.getItem('token');
 
-    const direcaoVotos = () => {
+    const votoPositive = () => {
+        if(curtida === 0){
+            curtida = 1
+        }else if(curtida === 1){
+            curtida =0
+        }
         const body = {
-            direction: count + 1
+            direction: curtida
         }
         axios.put(`https://us-central1-labenu-apis.cloudfunctions.net/labEddit/posts/${pathParams.idPost}/comment/${props.commentId}/vote`, body, {
                 headers: {
@@ -20,11 +25,33 @@ const CardComentario = (props) => {
                 }
             }).then(resposta => {
                 console.log(resposta.data);
+
             }).catch(error => {
                 console.error(error);
             })
-            
     };
+
+    const votoNegative = () => {
+        if(curtida === 0){
+            curtida = -1
+        }else if(curtida === -1){
+            curtida = 0
+        }
+        const body = {
+            direction: curtida
+        }
+        axios.put(`https://us-central1-labenu-apis.cloudfunctions.net/labEddit/posts/${pathParams.idPost}/comment/${props.commentId}/vote`, body, {
+                headers: {
+                    Authorization: token
+                }
+            }).then(resposta => {
+                console.log(resposta.data);
+
+            }).catch(error => {
+                console.error(error);
+            })
+    };
+
         return (
             <CardComentarios>
                 <Title>
@@ -33,7 +60,7 @@ const CardComentario = (props) => {
                 <TextPost>{props.text}</TextPost>
                 <BottomPost>
                     <Votos>
-                        <UpIcon onClick={direcaoVotos} />{props.votesCount}<DownIcon  />
+                        <UpIcon onClick={votoPositive} />{props.votesCount}<DownIcon onClick={votoNegative}/>
                     </Votos>
                 </BottomPost>
             </CardComentarios>
