+++ lab-eddit/src/pages/PagePost/index.js
@@ -2,34 +2,61 @@ import React, { useState, useEffect } from "react";
 import UpIcon from "@material-ui/icons/ArrowUpwardOutlined";
 import DownIcon from "@material-ui/icons/ArrowDownward";
 import { useParams } from "react-router-dom";
-import { useRequestDataGetDetail, useDataComments } from "../../CustomHooks/useRequestDataGet";
 import CardComentario from './CardComentario';
 import InputComentar from './InputComentar';
 import { ContainerPosts, CardPost, Title, TituloPost, TextPost, BottomPost, Votos, Comentarios } from '../styled';
-import { comentar } from '../../CustomHooks/useRequestDataPost';
-
+import {useHistory} from 'react-router-dom'
+import axios from 'axios'
 
 const PagePost = () => {
   const pathParams = useParams();
   const [texto, setTexto] = useState('');
+  const [post, setPost] = useState([])
+  const [comments, setComments] = useState([])
+  const history = useHistory();
+  const token = localStorage.getItem('token');
+
+  useEffect(()=>{
+    if (token === null) {
+        history.push("/");
+    }
+    atualizar()
+}, [history])
 
   const onChangeTexto = (ev) => {
     setTexto(ev.target.value);
   };
 
-  const post = useRequestDataGetDetail(
-    `https://us-central1-labenu-apis.cloudfunctions.net/labEddit/posts/${pathParams.idPost}`,
-    []
-  );
-  const comments = useDataComments(
-    `https://us-central1-labenu-apis.cloudfunctions.net/labEddit/posts/${pathParams.idPost}`,
-    []
-  );
+  const atualizar = ()=>{
+    axios.get(`https://us-central1-labenu-apis.cloudfunctions.net/labEddit/posts/${pathParams.idPost}`, {
+        headers:{
+            Authorization: token
+        }
+    }).then((response)=>{
+        console.log(response.data.post)
+        setPost(response.data.post)
+        setComments(response.data.post.comments)
+    }).catch((error)=>{
+        console.error(error)
+    })
+  }
+  
+
   const enviarComentario = () => {
     const body = {
       text: texto
     };
-    comentar(`https://us-central1-labenu-apis.cloudfunctions.net/labEddit/posts/${pathParams.idPost}/comment`, body);
+    axios.post(`https://us-central1-labenu-apis.cloudfunctions.net/labEddit/posts/${pathParams.idPost}/comment`, body, {
+      headers: {
+          Authorization: token
+      }
+    }).then(resposta => {
+        console.log(resposta.data);
+        atualizar()
+    }).catch(error => {
+        console.error(error);
+    });
+    
     setTexto('');
   };
